[dogstatsd] log useful parse errors
Previously dogstasd was only logging
```
            except Exception:
                log.exception('Error receiving datagram')
```
whereas the `_parse_*` functions were trying to raise useful errors.

This shouldn't happen anymore, as the parse exception are now handled and
logged in the aggregator.

It has another benefit: if there are incorrect packets in the "datagram",
the correct ones will be processed correctly (it was not the case before,
dogstatsd was failing at the first incorrect packet).