[http_check] use requests ssl_ instead of ssl
In order to get the certificate, `ssl_` is now used.
This was done to fix a sneaky bug which seems to only happen with some hosts
and not others. (only thing in common these hosts had was that they were
running Apache and returning a 401 for the queried URL)

`server_hostname` argument is needed, as this fix doesn't work otherwise.
According to the doc, it is set when the shipped OpenSSL version is
SNI-compatible.